LayoutTests/inspector/model/auditTestCase-expected.txt

Testing the functions of WI.AuditTestCase.


== Running test suite: AuditTestCase
-- Running test case: AuditTestCase.fromPayload.nullObject
null

-- Running test case: AuditTestCase.fromPayload.nonObject
null

-- Running test case: AuditTestCase.fromPayload.emptyObject
null

-- Running test case: AuditTestCase.fromPayload.invalidTopLevelMembers
null

-- Running test case: AuditTestCase.fromPayload.valid
{
  "type": "test-case",
  "name": "valid test name",
  "test": "function() { }"
}

-- Running test case: AuditTestCase.fromPayload.validWithInvalidOptionals
WARN: Audit Warning: "validWithInvalidOptionals test name" has a non-string "description" value
WARN: Audit Warning: "validWithInvalidOptionals test name" has a non-string "setup" value
WARN: Audit Warning: "validWithInvalidOptionals test name" is too new to run in this Web Inspector
{
  "type": "test-case",
  "name": "validWithInvalidOptionals test name",
  "supports": 3,
  "test": "validWithInvalidOptionals test function"
}

-- Running test case: AuditTestCase.fromPayload.validWithValidOptionals
{
  "type": "test-case",
  "name": "validWithValidOptionals test name",
  "description": "validWithValidOptionals test description",
  "supports": 1,
  "setup": "validWithValidOptionals test setup",
  "test": "validWithValidOptionals test function"
}
